+++ alphadev/data/loaders/__init__.py
@@ -4,6 +4,7 @@ from .base import DataLoader
 from .alpha_loader import AlphaLoader, AlphaRankLoader, DEFAULT_ALPHA_DIR
 from .common_alphas_loader import CommonAlphasLoader, DEFAULT_COMMON_ALPHAS_DIR
 from .feature_loader import FeatureLoader
+from .universe_loader import UniverseLoader
 from .csv_loader import CSVDataLoader
 from .composite import CompositeDataLoader
 from .kline import KlineDataLoader
@@ -17,6 +18,7 @@ __all__ = [
     "CommonAlphasLoader",
     "DEFAULT_COMMON_ALPHAS_DIR",
     "FeatureLoader",
+    "UniverseLoader",
     "CSVDataLoader",
     "CompositeDataLoader",
     "KlineDataLoader",